Update chapter2_content.tex
My understanding is that A2 doesn't need to be exploratory since it could be exploiting the initial expected value of 0 after getting r=-1 reward from A1. 
Also, A4 must be explorative because the choice is for a=2 when at that point the estimated value is Q3(a=2)=-0.5, while Q3(a=3)=0, Q3(a=4)=0.